+++ crates/schematic/src/config/parser.rs
@@ -1,8 +1,9 @@
-use miette::{Diagnostic, SourceSpan};
+use miette::{Diagnostic, NamedSource, SourceSpan};
 use starbase_styles::{Style, Stylize};
+use std::borrow::Borrow;
 use thiserror::Error;
 
-/// Error for a single parse failure. Can be used with [`TryFrom`], [`FromStr`], or serde.
+/// Error for a single parse failure.
 #[derive(Clone, Debug, Diagnostic, Error)]
 #[error("{message}")]
 pub struct ParseError {
@@ -20,17 +21,27 @@ impl ParseError {
 }
 
 /// Error related to serde parsing.
-#[derive(Error, Debug, Diagnostic)]
+#[derive(Debug, Diagnostic, Error)]
 #[error("{}{} {message}", .path.style(Style::Id), ":".style(Style::MutedLight))]
 #[diagnostic(severity(Error))]
 pub struct ParserError {
+    /// Source code snippet related to the error.
     #[source_code]
-    pub content: String, // NamedSource,
+    pub content: NamedSource<String>,
 
+    /// Failure message.
     pub message: String,
 
+    /// Dot-notated path to the field that failed.
     pub path: String,
 
+    /// Span to the error location.
     #[label("Fix this")]
     pub span: Option<SourceSpan>,
+}
+
+impl Borrow<dyn Diagnostic> for Box<ParserError> {
+    fn borrow(&self) -> &(dyn Diagnostic + 'static) {
+        self.as_ref()
+    }
 }