+++ crates/schematic/src/config/source.rs
@@ -1,6 +1,6 @@
-use crate::config::cacher::BoxedCacher;
-use crate::config::errors::ConfigError;
-use crate::config::format::Format;
+use super::cacher::BoxedCacher;
+use super::error::ConfigError;
+use crate::format::Format;
 use serde::Deserialize;
 use serde::{de::DeserializeOwned, Serialize};
 use std::fs;
@@ -46,25 +46,22 @@ impl Source {
                 value
             };
 
-            if parent_source.is_none() {
-                return Source::file(value, true);
-            }
+            return match parent_source {
+                None => Source::file(value, true),
+                Some(Source::File {
+                    path: parent_path, ..
+                }) => {
+                    let mut path = PathBuf::from(value);
 
-            if let Source::File {
-                path: parent_path, ..
-            } = parent_source.unwrap()
-            {
-                let mut path = PathBuf::from(value);
+                    // Not absolute, so prefix with parent
+                    if !path.has_root() {
+                        path = parent_path.parent().unwrap().join(path);
+                    }
 
-                // Not absolute, so prefix with parent
-                if !path.has_root() {
-                    path = parent_path.parent().unwrap().join(path);
+                    Source::file(path, true)
                 }
-
-                return Source::file(path, true);
-            } else {
-                return Err(ConfigError::ExtendsFromParentFileOnly);
-            }
+                Some(_) => Err(ConfigError::ExtendsFromParentFileOnly),
+            };
         }
 
         Err(ConfigError::ExtendsFromNoCode)
@@ -99,26 +96,19 @@ impl Source {
     }
 
     /// Parse the source contents according to the required format.
-    #[instrument(name = "parse_config_source", skip(cacher, help), fields(source = ?self))]
-    pub fn parse<D>(
-        &self,
-        location: &str,
-        cacher: &mut BoxedCacher,
-        help: Option<&str>,
-    ) -> Result<D, ConfigError>
+    #[instrument(name = "parse_config_source", skip(cacher), fields(source = ?self))]
+    pub fn parse<D>(&self, name: &str, cacher: &mut BoxedCacher) -> Result<D, ConfigError>
     where
         D: DeserializeOwned,
     {
-        let handle_error = |error: crate::config::ParserError| ConfigError::Parser {
-            config: location.to_owned(),
-            error,
-            help: help.map(|h| h.to_owned()),
+        let handle_error = |error: super::parser::ParserError| ConfigError::Parser {
+            location: String::new(),
+            error: Box::new(error),
+            help: None,
         };
 
         match self {
-            Source::Code { code, format } => format
-                .parse(code.to_owned(), location)
-                .map_err(handle_error),
+            Source::Code { code, format } => format.parse(name, code).map_err(handle_error),
             Source::File {
                 path,
                 format,
@@ -137,7 +127,7 @@ impl Source {
                     "".into()
                 };
 
-                format.parse(content, location).map_err(handle_error)
+                format.parse(name, &content).map_err(handle_error)
             }
             Source::Url { url, format } => {
                 if !is_secure_url(url) {
@@ -164,7 +154,7 @@ impl Source {
                         body
                     };
 
-                    format.parse(content, location).map_err(handle_error)
+                    format.parse(name, &content).map_err(handle_error)
                 }
 
                 #[cfg(not(feature = "url"))]